requirements.txt
streamlit>=1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
edge-tts>=6.1.0
supabase>=2.0.0
